@@ -7,7 +7,7 @@ class LikesController < ApplicationController
     @like = current_user.likes.new(post_id: params[:post_id])
 
     if @like.save
-      redirect_to posts_path, notice: 'You liked a post.'
+      redirect_to posts_path, notice: 'You liked this post.'
     else
       redirect_to posts_path, alert: 'You cannot like this post.'
     end
