@@ -10,7 +10,7 @@ class CreateFriendships < ActiveRecord::Migration[6.1]
 
       t.timestamps
     end
-
+    
   end
 end
 # rubocop:enable Style/Documentation
